fix(backend): moved the person files to the test folder
