+++ power-pay-backend/src/test/java/com/adorsys/gis/powerpay/powerpaybackend/Person.java
@@ -1,4 +1,4 @@
-package com.adorsys.gis.powerpay.powerpaybackend.domain.model;
+package com.adorsys.gis.powerpay.powerpaybackend;
 
 public class Person {
     private String name;
