+++ power-pay-backend/src/test/java/com/adorsys/gis/powerpay/powerpaybackend/PersonDto.java
@@ -1,4 +1,4 @@
-package com.adorsys.gis.powerpay.powerpaybackend.services.dto;
+package com.adorsys.gis.powerpay.powerpaybackend;
 
 // PersonDto.java
 public class PersonDto {
