+++ power-pay-backend/src/test/java/com/adorsys/gis/powerpay/powerpaybackend/PersonMapper.java
@@ -1,4 +1,4 @@
-package com.adorsys.gis.powerpay.powerpaybackend.utils;
+package com.adorsys.gis.powerpay.powerpaybackend;
 
 // PersonMapper.java
 import org.mapstruct.Mapper;
